+++ spec/models/composite.rb
@@ -2,4 +2,5 @@
 
 class Composite < Sequel::Model(:composite_pks)
   unrestrict_primary_key
+  one_to_many :composite_children, key: [:a, :b]
 end