+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.5.0
+- Add: OneToMany::Plural::Create
+
 ## 0.4.6
 - Fix: typos/wording in README